Move update to API module
